*)reverted patch for memory-display issue
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2095 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/PerformanceMemory_p.java
@@ -63,7 +63,7 @@ public class PerformanceMemory_p {
     private static final long MB = 1024 * KB;
     private static Map defaultSettings = null;
         
-    private static long[]    slt,chk;
+    private static int[]    slt,chk;
     private static String[] ost;
     private static long     req, usd, bst, god;
